Improve the parsing of author lines from Mercurial to generate committer and author lines that git can correctly understand
Signed-off-by: Ehsan Akhgari <ehsan.akhgari@gmail.com>
---
I found a number of bugs when I was trying to convert Mozila's hg repository
to git using hg-git.  This patch fixes a number of bugs with irregular
author lines present in hg repositories.  Git cannot correctly process a
commit object which has a committer or author line in a format that it does
not understand, which makes it not be able to handle the repositories
with have such commit objects.

The added test cases shows the irregular cases that this patch is able to
deal with.