tests: remove piping of command git init to remove leading path
In many cases we were piping to a python script to strip out the varying leading
path to the test repos.  This is no longer needed, as the modern run-test.py
automatically substitutes the leading path as $TESTTMP.  Eliminating the piping
makes the tests easier to read and write, as well as allowing the exit codes
to be verified by the test.