push: fix exchangepush wrapping
Upstream added opargs to exchange.push and uses it as kwargs to the
pushoperation constructor (16e69e6b357b). There was an attempt to fix this in
hggit (a98b3a70edc3) but it passes the exchange.push kwargs directly to
pushoperation(), where we actually need to pull out the opargs and pass them as
kwargs.